:truck: moved cover component to app module
