+++ frontend/src/modules/app/components/cover/index.js
@@ -4,8 +4,8 @@ import { Link } from "react-router-dom";
 import { Image } from "semantic-ui-react";
 import "./styles.css";
 
-const Cover = ({ imageId, slug, className }) => {
-  const cover = `https://images.igdb.com/igdb/image/upload/t_cover_big/${imageId}.jpg`;
+const Cover = ({ imageId, slug, className, size }) => {
+  const cover = `https://images.igdb.com/igdb/image/upload/t_cover_${size}/${imageId}.jpg`;
   return (
     <div className={className}>
       <Link to={`/games/${slug}`}>
