+++ frontend/src/modules/app/components/cover/styles.css
