+++ frontend/src/modules/game/components/index.js
@@ -1,5 +1,4 @@
 export { default as Actions } from "./actions";
-export { default as Cover } from "./cover";
 export { default as Details } from "./details";
 export { default as TitleLoader } from "./loaders/TitleLoader";
 export { default as CoverLoader } from "./loaders/CoverLoader";
